Moved TransactionException and TransactionExtensions out of samples and into the main library, as they are generally useful.
