+++ TA.Ascom.ReactiveCommunications/TA.Ascom.ReactiveCommunications.csproj
@@ -24,6 +24,7 @@ For more information and a quick-start tutorial, please see http://tigra-astrono
   </PropertyGroup>
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
+    <LangVersion>latest</LangVersion>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <PackageProjectUrl>https://github.com/Tigra-Astronomy/TA.ReactiveCommunications/wiki</PackageProjectUrl>
     <PackageIconUrl>https://github.com/Tigra-Astronomy/TA.ReactiveCommunications/wiki/Artwork/NugetPackageIcon.png</PackageIconUrl>
